fix: 육류 데이터 관리 일반데이터 승인 페이지 palette, margin  완료
