+++ app/structure/lib/screen/data_management/normal/edit_meat_data_screen.dart
@@ -10,6 +10,7 @@ import 'package:structure/components/custom_app_bar.dart';
 import 'package:structure/components/loading_screen.dart';
 import 'package:structure/components/round_button.dart';
 import 'package:structure/components/step_card.dart';
+import 'package:structure/config/palette.dart';
 import 'package:structure/config/pallete.dart';
 import 'package:structure/viewModel/data_management/normal/edit_meat_data_view_model.dart';
 
@@ -22,88 +23,101 @@ class EditMeatDataScreen extends StatelessWidget {
         context.watch<EditMeatDataViewModel>();
 
     return Scaffold(
-      backgroundColor: Colors.white,
       appBar: CustomAppBar(
-          title: '${editMeatDataViewModel.meatModel.meatId}',
-          backButton: true,
-          closeButton: false),
+        title: '${editMeatDataViewModel.meatModel.meatId}',
+        backButton: true,
+        closeButton: false,
+      ),
       body: Stack(
         children: [
-          Column(
-            children: [
-              SizedBox(height: 48.h),
+          Container(
+            margin: EdgeInsets.symmetric(horizontal: 40.w),
+            child: Column(
+              children: [
+                SizedBox(height: 24.h),
 
-              // 육류 기본 정보
-              StepCard(
-                mainText: '육류 기본정보',
-                status: editMeatDataViewModel.isNormal
-                    ? editMeatDataViewModel.isEditable
-                        ? 3 // 수정 가능
-                        : 4 // 수정 불가
-                    : null, // 없음
+                // 육류 기본 정보
+                StepCard(
+                  mainText: '육류 기본정보',
+                  status: editMeatDataViewModel.isNormal
+                      ? editMeatDataViewModel.isEditable
+                          ? 3 // 수정 가능
+                          : 4 // 수정 불가
+                      : null, // 없음
 
-                onTap: () => editMeatDataViewModel.clicekdBasic(context),
-                imageUrl: 'assets/images/meat_info.png',
-              ),
-              SizedBox(height: 18.h),
+                  onTap: () => editMeatDataViewModel.clicekdBasic(context),
+                  imageUrl: 'assets/images/meat_info.png',
+                ),
+                SizedBox(height: 16.h),
 
-              // 육류 단면 촬영
-              StepCard(
-                mainText: '육류 단면 촬영',
-                status: editMeatDataViewModel.isNormal
-                    ? editMeatDataViewModel.isEditable
-                        ? 3 // 수정 가능
-                        : 4 // 수정 불가
-                    : null, // 없음
-                onTap: () => editMeatDataViewModel.clickedImage(context),
-                imageUrl: 'assets/images/meat_image.png',
-              ),
-
-              SizedBox(height: 18.h),
-
-              // 원육 관능 평가
-              StepCard(
-                mainText: '원육 관능평가',
-                status: editMeatDataViewModel.isNormal
-                    ? editMeatDataViewModel.isEditable
-                        ? 3 // 수정 가능
-                        : 4 // 수정 불가
-                    : null, // 없음
-                onTap: () => editMeatDataViewModel.clicekdFresh(context),
-                imageUrl: 'assets/images/meat_eval.png',
-              ),
-
-              const Spacer(),
-
-              // 연구자 신분 + 승인되지 않은 데이터일때만
-              if (editMeatDataViewModel.showAcceptBtn())
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    RoundButton(
-                      onPress: () {
-                        editMeatDataViewModel.rejectMeatData(context);
-                      },
-                      text: Text('반려', style: Pallete.fieldPlaceHolderWhite),
-                      bgColor: Pallete.alertColor,
-                      width: 310.w,
-                      height: 96.h,
-                    ),
-                    SizedBox(width: 20.w),
-                    RoundButton(
-                      onPress: () {
-                        editMeatDataViewModel.acceptMeatData(context);
-                      },
-                      text: Text('승인', style: Pallete.fieldPlaceHolderWhite),
-                      bgColor: Pallete.checkSpeciesColor,
-                      width: 310.w,
-                      height: 96.h,
-                    ),
-                  ],
+                // 육류 단면 촬영
+                StepCard(
+                  mainText: '육류 단면 촬영',
+                  status: editMeatDataViewModel.isNormal
+                      ? editMeatDataViewModel.isEditable
+                          ? 3 // 수정 가능
+                          : 4 // 수정 불가
+                      : null, // 없음
+                  onTap: () => editMeatDataViewModel.clickedImage(context),
+                  imageUrl: 'assets/images/meat_image.png',
                 ),
 
-              SizedBox(height: 40.h),
-            ],
+                SizedBox(height: 16.h),
+
+                // 원육 관능 평가
+                StepCard(
+                  mainText: '원육 관능평가',
+                  status: editMeatDataViewModel.isNormal
+                      ? editMeatDataViewModel.isEditable
+                          ? 3 // 수정 가능
+                          : 4 // 수정 불가
+                      : null, // 없음
+                  onTap: () => editMeatDataViewModel.clicekdFresh(context),
+                  imageUrl: 'assets/images/meat_eval.png',
+                ),
+
+                const Spacer(),
+
+                // 연구자 신분 + 승인되지 않은 데이터일때만
+                if (editMeatDataViewModel.showAcceptBtn())
+                  Container(
+                    margin: EdgeInsets.only(bottom: 40.h),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        // 반려 버튼
+                        Expanded(
+                          child: RoundButton(
+                            onPress: () {
+                              editMeatDataViewModel.rejectMeatData(context);
+                            },
+                            text: Text('반려',
+                                style: Pallete.fieldPlaceHolderWhite),
+                            bgColor: Palette.error,
+                            width: double.infinity,
+                            height: 96.h,
+                          ),
+                        ),
+                        SizedBox(width: 16.w),
+
+                        // 승인 버튼
+                        Expanded(
+                          child: RoundButton(
+                            onPress: () {
+                              editMeatDataViewModel.acceptMeatData(context);
+                            },
+                            text: Text('승인',
+                                style: Pallete.fieldPlaceHolderWhite),
+                            bgColor: Palette.primary,
+                            width: double.infinity,
+                            height: 96.h,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+              ],
+            ),
           ),
           editMeatDataViewModel.isLoading
               ? const Center(child: LoadingScreen())
